(edit) update the helpers/search/*
